Accurate values from widgets without change events! See description, it's not a perfect fix
I've hijacked how ooba webui gets its element values by injecting the BHC elements into it. The result is that `state` is now generated with the BHC element values with perfect accuracy. I then take the values out of `state` and turn them into `ui_params` for `chip.py`, which then turns that into the `param` of `brainhackingchip_settings` for each chip when called. So by accessing `param` in `brainhackingchip_settings`, you get what should be accurate values for all widgets for that chip.

The structure is similar to what `traverse_to` expects but won't work with it, it's the same structure I was using for the save file. Each chip's `param` is a dict containing keys for widgets. Each value in the dict is another dict, that may have a `'value'` (which is the value) or `'sub'` (which are the sub_attr, another dict). The structure is nested, with `'sub'` being a dict with more keys that then follow this pattern.

If all this UI hooking is an issue for any reason, it can probably be fixed by commenting out or renaming the `setup()` function in `script.py`.

Problems:
- After the initial ui() call, BHC widgets can't be added or removed. Currently, all widgets from every viable chip are loaded so that's fine, but attempting to add a new chip to the directory and refresh the widgets without actually reloading ooba will probably break things. I'd like to fix this, but it isn't easy.
- Session tab is broken. I removed it because it will currently show up inside the BHC tab because I had to change the order that various UI things are initialized. Maybe there's a clever way to be able to put it back in its place, the commented `# func()` on 392 of `script.py` is the call to create the session tab.