Fixed default to use ui_params
While I can get values in the GUI and ui_params to desync by holding down the arrow key on a slider sometimes, I'm not sure if that issue justifies my workaround. Any other situation that causes the change event to be rapidly called will probably have the same problem, and I'm not sure if slower systems will have this issue worse. The release event seems to fix the major case causing ui_params to have values not matching the GUI though, so hopefully it is enough?